+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - Tests and coverage GitHub workflow
 - Updated submission schema in resources following new release of schema by ClinVar (draft-07)
 - Proxy endpoint to interrogate the submission dry run endpoint of ClinVar API
+- Proxy endpoint to interrogate the submission validate endpoint of ClinVar API